requirements.txt
Flask
torch==2.2.0 --index-url https://download.pytorch.org/whl/cpu
torchaudio==2.2.0 --index-url https://download.pytorch.org/whl/cpu
transformers==4.37.2
PyPDF2
langchain
langchain-google-genai
soundfile
numpy
requests
python-dotenv